# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Agent Platform API
  description: |
    Agent Platform Product Experience API - A multi-agent platform for task orchestration 
    and execution with LLM providers, background task processing, workflows, triggers, and connectors.
    
    All endpoints use a standard API response envelope:
    - Success: `{ ok: true, data: ..., requestId: "..." }`
    - Error: `{ ok: false, error: { code: "...", message: "..." }, requestId: "..." }`
  version: "0.7.0-rc.1"
  contact:
    name: Agent Platform Team
servers:
  - url: http://localhost:3003
    description: Development
  - url: http://localhost:3103
    description: E2E Testing
...
paths:
...
  /api/v1/api-keys:
    get:
      tags: [RBAC]
      summary: List API keys
      description: Returns all API keys for the current user
      operationId: listApiKeys
      security:
        - sessionAuth: []
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [RBAC]
      summary: Create API key
      description: Creates a new API key for programmatic access
      operationId: createApiKey
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiKeyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/api-keys/{keyId}:
    delete:
      tags: [RBAC]
      summary: Revoke API key
      description: Revokes (deletes) an API key
      operationId: revokeApiKey
      security:
        - sessionAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: API key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeApiKeyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==========================================================================
  # Sessions
  # ==========================================================================
...
  /api/v1/debug/replay/{sessionId}:
    get:
      tags: [System]
      summary: Debug replay
      description: Returns debug replay information for a session
      operationId: debugReplay
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/sessionId"
      responses:
        "200":
          description: Replay debug info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugReplayResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

# ============================================================================
# Component Schemas
# ============================================================================
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: agent-platform-session
      description: Session cookie set after login
    bearerAuth:
      type: http
      scheme: bearer
      description: API token for programmatic access (set API_AUTH_TOKEN env var to enable)
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session ID
...
  schemas:
...
    ApiEnvelope:
      type: object
      required: [ok, requestId]
      properties:
        ok:
          type: boolean
          description: true for success, false for error
        requestId:
          type: string
          description: Traceable request identifier
    ApiErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          description: Additional error details
...
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [error]
          properties:
            error:
              $ref: "#/components/schemas/ApiErrorDetail"

    # ---- Auth & Setup ----
...
    ApiKeyListEntry:
      type: object
      required: [id, name, prefix, role, userId, createdAt, isActive]
      properties:
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
          description: First 8 characters of the key for identification
        role:
          type: string
          enum: [admin, user, service]
        userId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
    ApiKeyListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ApiKeyListEntry"
    CreateApiKeyRequest:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [admin, user, service]
        expiresAt:
          type: string
          format: date-time
    CreateApiKeyResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [id, name, key, prefix, role, createdAt]
              properties:
                id:
                  type: string
                name:
                  type: string
                key:
                  type: string
                  description: The full API key (only shown once on creation)
                prefix:
                  type: string
                role:
                  type: string
                createdAt:
                  type: string
                  format: date-time
    RevokeApiKeyResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [id, isActive]
              properties:
                id:
                  type: string
                isActive:
                  type: boolean

    # ---- Alerting ----
...
    DebugReplayResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                eventCount:
                  type: integer
                transcriptCount:
                  type: integer
                runRefs:
                  type: array
                  items:
                    type: string
                approvalRefs:
                  type: array
                  items:
                    type: string
                lastEventId:
                  type: string
                redactedPreviews:
                  type: array
                  items:
                    type: object

    # ---- Shared ----
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "BAD_REQUEST"
              message: "Invalid request body"
            requestId: "abc12345"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
            requestId: "abc12345"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
            requestId: "abc12345"
...